add some comments
