+++ src/OpenRiaServices.Client.DomainClients.Http/Framework/Http/BinaryHttpDomainClient.cs
@@ -1,9 +1,14 @@
 ﻿using System;
 using System.IO;
 using System.Net.Http;
+using System.Runtime.Serialization;
 
 namespace OpenRiaServices.Client.DomainClients.Http
 {
+    /// <summary>
+    /// <see cref="DomainClient"/> implementation which uses plain binary Xml using <see cref="DataContractSerializer"/> using the default format
+    /// since WCF RIA Services.
+    /// </summary>
     sealed class BinaryHttpDomainClient : DataContractHttpDomainClient
     {
         internal const string MediaType = "application/msbin1";
@@ -14,6 +19,7 @@ namespace OpenRiaServices.Client.DomainClients.Http
 
         private protected override string ContentType => MediaType;
 
+
         private protected override System.Xml.XmlDictionaryReader CreateReader(Stream stream)
         {
             return System.Xml.XmlDictionaryReader.CreateBinaryReader(stream, System.Xml.XmlDictionaryReaderQuotas.Max);
@@ -22,29 +28,6 @@ namespace OpenRiaServices.Client.DomainClients.Http
         private protected override System.Xml.XmlDictionaryWriter CreateWriter(Stream stream)
         {
             return System.Xml.XmlDictionaryWriter.CreateBinaryWriter(stream, null, null, ownsStream: false);
-        }
-    }
-
-    sealed class XmlHttpDomainClient : DataContractHttpDomainClient
-    {
-        internal const string MediaType = "application/xml";
-
-        private readonly System.Text.Encoding _encoding = new System.Text.UTF8Encoding(encoderShouldEmitUTF8Identifier: false);
-
-        public XmlHttpDomainClient(HttpClient httpClient, Type serviceInterface) : base(httpClient, serviceInterface)
-        {
-        }
-
-        private protected override string ContentType => MediaType;
-
-        private protected override System.Xml.XmlDictionaryReader CreateReader(Stream stream)
-        {
-            return System.Xml.XmlDictionaryReader.CreateTextReader(stream, System.Xml.XmlDictionaryReaderQuotas.Max);
-        }
-
-        private protected override System.Xml.XmlDictionaryWriter CreateWriter(Stream stream)
-        {
-            return System.Xml.XmlDictionaryWriter.CreateTextWriter(stream, _encoding, ownsStream: false);
         }
     }
 }
